Define Policy for delaying dispatcher events
This allows the programmer to specify a delay in seconds (and fractions
of a second) as a function of some to-be-determined statistics, as well
as a random number generator and the current time (which may be useful
alongside the statistics, which will also include absolute times).

Some changes to the examples have also been included: both of them now
create a new TCP transport for each of the nodes (previously they all
shared the same transport, and so their dispatcher threads all blocked
on the same incoming event). Both have also been given dispatcher delay
policies, but to different results:

  PingPong works fine because the worker thread blocks until the peer
  responds, thus the system proceeds synchronously: ping, then wait for
  pong, then repeat.

  Discovery does *not* work fine because it has been defined so that the
  worker will proceed even as pings are in-flight but unacknowledged.
  So a slow node (simulated by a delay policy) will cause its naive
  peers to overload themselves, repeatedly trying to ping it even though
  the messages are not getting through.